part 1

+++ be/src/vec/columns/predicate_column.h
@@ -32,6 +32,7 @@
 #include "vec/core/types.h"
 
 namespace doris::vectorized {
+#include "common/compile_check_begin.h"
 
 /**
  * used to keep predicate column in storage layer
@@ -285,7 +286,7 @@ public:
             char* org_dst = destination;
             size_t org_elem_num = data.size();
             data.resize(org_elem_num + num);
-            uint32_t fragment_start_offset = 0;
+            uint64_t fragment_start_offset = 0;
             size_t fragment_len = 0;
             for (size_t i = 0; i < num; i++) {
                 data[org_elem_num + i].data = destination + fragment_len;
@@ -440,4 +441,5 @@ private:
     std::vector<StringRef> _refs;
 };
 
+#include "common/compile_check_end.h"
 } // namespace doris::vectorized
